+++ packages/hydrogen/CHANGELOG.md
@@ -43,7 +43,7 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 - Remove the `<SelectedVariantImage />` component in favour of using `<Image data={product.selectedVariant.image} />`
 - Remove the `<SelectedVariantMetafield />` component in favour of using `<ProductMetafield variantId={product.selectedVariant.id} />`
 - Remove the `<SelectedVariantShopPayButton />` component in favour of using `<ShopPayButton variantId={product.selectedVariant.id} />`
-- Remove the `<SelectedVariantPrice/>` component in favour of using `<ProductPrice variantId={product.selectedVariant.id} />`
+- Remove the `<SelectedVariantPrice/>` and `<SelectedVariantUnitPrice/>` component in favour of using `<ProductPrice variantId={product.selectedVariant.id} />`
 
 ### Changed